simplify throwing entity effects
