--- Content.Shared/_CE/EntityEffect/Effects/ThrowFromUser.cs
@@ -1,41 +1,0 @@
-﻿using System.Numerics;
-using Content.Shared.Projectiles;
-using Content.Shared.Throwing;
-
-namespace Content.Shared._CE.EntityEffect.Effects;
-
-public sealed partial class ThrowFromUser : CEEntityEffectBase<ThrowFromUser>
-{
-    [DataField]
-    public float ThrowPower = 10f;
-
-    [DataField]
-    public float Distance = 2.5f;
-}
-
-public sealed partial class CEThrowFromUserEffectSystem : CEEntityEffectSystem<ThrowFromUser>
-{
-    [Dependency] private ThrowingSystem _throwing = default!;
-    [Dependency] private SharedTransformSystem _transform = default!;
-    [Dependency] private SharedProjectileSystem _projectile = default!;
-
-    protected override void Effect(ref CEEntityEffectEvent<ThrowFromUser> args)
-    {
-        if (ResolveEffectEntity(args.Args, args.Effect.EffectTarget) is not { } targetEntity)
-            return;
-
-        var worldPos = _transform.GetWorldPosition(args.Args.Source);
-        var dir = _transform.GetWorldPosition(targetEntity) - worldPos;
-        if (dir == Vector2.Zero)
-            return;
-
-        var normalized = Vector2.Normalize(dir);
-
-        if (TryComp<EmbeddableProjectileComponent>(targetEntity, out var embeddable))
-        {
-            _projectile.EmbedDetach(targetEntity, embeddable);
-        }
-
-        _throwing.TryThrow(targetEntity, normalized * (args.Effect.Distance * args.Args.Power), args.Effect.ThrowPower * args.Args.Power, doSpin: true);
-    }
-}
