+++ Resources/Prototypes/_CE/Entities/Mobs/NPCs/slime_water_big.yml
@@ -1,4 +1,4 @@
-- type: entity
+﻿- type: entity
   parent:
   - BaseMob
   - CEMobDamageable
@@ -58,7 +58,7 @@
     - !type:AreaEffect
       range: 1
       effects:
-      - !type:ThrowFromUser
+      - !type:ThrowFrom
         distance: 1
   - type: CEGOAP
 
@@ -111,7 +111,7 @@
           damage:
             types:
               Physical: 15
-        - !type:ThrowFromUser
+        - !type:ThrowFrom
           distance: 1
     animations:
       Primary:
