+++ Resources/Prototypes/_CE/Skills/Knight/Actives/kick.yml
@@ -1,4 +1,4 @@
-- type: skill
+﻿- type: skill
   parent: BaseKnightSkill
   id: Kick
   effect: !type:AddAction
@@ -72,7 +72,7 @@
         damage:
           types:
             Physical: 15
-      - !type:ThrowFromUser
+      - !type:ThrowFrom
         distance: 3
         throwPower: 10
